requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
python-dotenv>=1.0.0
mcp[cli]>=1.0.0
pdfplumber>=0.10.0
sentence-transformers>=2.7.0
numpy>=1.24.0

# RAGAS evaluation (dev-only — not needed at runtime)
ragas>=0.2.0
datasets>=2.16.0
langchain-anthropic>=0.2.0
langchain-huggingface>=0.1.0
